Renamed `hint_processor.rs` to `calls.rs` (#588)
